Add support for Ruby w/o Rails.
