+++ Appraisals
@@ -1,0 +1,11 @@
+appraise 'rails_4' do
+  gem 'rails', '~> 4.0'
+end
+
+appraise 'rails_5.0' do
+  gem 'rails', '~> 5.0.0'
+end
+
+appraise 'rails_5.1' do
+  gem 'rails', '~> 5.1.0'
+end
