+++ CHANGELOG.md
@@ -1,10 +1,23 @@
+## 5.0.0 (Next)
+
+* [#103](https://github.com/radar/distance_of_time_in_words/pull/103): Support for Ruby w/o Rails - [@dblock](https://github.com/dblock).
+
+## 4.0.1 (2018/06/01)
+
 ## 3.1.1 (2016/03/08)
-* Add CHANGELOG.md
 
 ## 3.1.0 (2016/03/07)
 
-* [#68](https://github.com/radar/dotiw/pull/68): Add support for weeks - [@lauranjansen](https://github.com/lauranjansen)
-* [#69](https://github.com/radar/dotiw/pull/69): Add Indonesian language support - [@avidmaulanas](https://github.com/avidmaulanas)
-* [#70](https://github.com/radar/dotiw/pull/70): Add French language support - [@geo1004](https://github.com/geo1004)
-* [#72](https://github.com/radar/dotiw/pull/72): Add Danish language support - [@kaspernj](https://github.com/kaspernj)
-* [#71](https://github.com/radar/dotiw/pull/71): Update bundler and ruby versions for Travis CI - [@lauranjansen](https://github.com/lauranjansen)
+* [#68](https://github.com/radar/dotiw/pull/68): Add support for weeks - [@lauranjansen](https://github.com/lauranjansen).
+* [#69](https://github.com/radar/dotiw/pull/69): Add Indonesian language support - [@avidmaulanas](https://github.com/avidmaulanas).
+* [#70](https://github.com/radar/dotiw/pull/70): Add French language support - [@geo1004](https://github.com/geo1004).
+* [#72](https://github.com/radar/dotiw/pull/72): Add Danish language support - [@kaspernj](https://github.com/kaspernj).
+* [#71](https://github.com/radar/dotiw/pull/71): Update bundler and ruby versions for Travis CI - [@lauranjansen](https://github.com/lauranjansen).
+
+## 3.0.1 (2015/04/09)
+
+## 3.0 (2015/04/09)
+
+## 0.2.4 (2009/10/18)
+
+* Initial public release - [@radar](https://github.com/radar).
