+++ MIT-LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2009 Ryan Bigg
+Copyright (c) 2009-2020 Ryan Bigg
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
