+++ gemfiles/rails_4.gemfile
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 # This file was generated by Appraisal
 
-source "http://rubygems.org"
+source 'http://rubygems.org'
 
-gem "rails", "~> 4.0"
+gem 'rails', '~> 4.0'
 
-gemspec path: "../"
+gemspec path: '../'
