+++ gemfiles/rails_5.0.gemfile
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 # This file was generated by Appraisal
 
-source "http://rubygems.org"
+source 'http://rubygems.org'
 
-gem "rails", "~> 5.0.0"
+gem 'rails', '~> 5.0.0'
 
-gemspec path: "../"
+gemspec path: '../'
