+++ lib/dotiw/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module DOTIW
-  VERSION = '4.0.1'.freeze
+  VERSION = '5.0.0'
 end
