+++ spec/spec_helper.rb
@@ -1,11 +1,7 @@
+# frozen_string_literal: true
+
 ROOT_PATH = File.join(File.dirname(__FILE__), '..')
 $LOAD_PATH.unshift ROOT_PATH unless $LOAD_PATH.include? ROOT_PATH
-
-# Files that are usually required by Rails, but in a testing context will not be.
-require 'erb'
-
-require 'active_support/all'
-require 'action_view'
 
 require 'dotiw'
